requirements.txt
# Core Framework
streamlit>=1.28.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# AWS
boto3>=1.34.0

# AI/ML
langgraph>=0.0.40
langchain>=0.1.0
anthropic>=0.40.0

# Databases
pinecone>=8.0.0
pymongo>=4.6.0
motor>=3.3.0

# PDF Generation
reportlab>=4.0.0
matplotlib>=3.8.0
pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
python-multipart>=0.0.6
